python/2019-10-07_Virtual_Assistant_v2/requirements.txt
SpeechRecognition
beautifulsoup4
python-vlc
youtube-dl
pyowm
wikipedia
